chore(scanjob): move legacy views to a dedicated module
