+++ quipucords/api/views.py
@@ -11,7 +11,8 @@ from api.insights_report.view import insights
 from api.merge_report.view import async_merge_reports
 from api.report.view_v1 import reports, upload_raw_facts
 from api.scan.view import ScanViewSet, jobs
-from api.scanjob.view import ScanJobViewSetV1, ScanJobViewSetV2
+from api.scanjob.view import ScanJobViewSet
+from api.scanjob.view_v1 import ScanJobViewSet as ScanJobViewSetV1
 from api.source.view import SourceViewSet, source_bulk_delete
 from api.status.view import ping, status
 from api.user.token_view import QuipucordsExpiringAuthTokenView
