+++ quipucords/tests/api/scanjob/test_scanjob.py
@@ -20,7 +20,7 @@ from api.models import (
 )
 from api.scan.serializer import ScanSerializer
 from api.scanjob.serializer import ScanJobSerializerV1
-from api.scanjob.view import expand_scanjob
+from api.scanjob.view_v1 import expand_scanjob
 from tests.api.scan.test_scan import (
     disabled_optional_products_default,
     enabled_extended_product_search_default,
